# openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Reactor
  version: 1.26.10-beta8
  license:
    name: BUSL-1.1
    url: https://mariadb.com/bsl11/
  description: |
    Auth, storage, functions, sites, health, and platform routes.
    Data stays on PostgREST at /data/v1.
servers:
  - url: http://127.0.0.1:18000
    description: Local Compose
...
paths:
...
  /auth/v1/magic-link:
    post:
      tags: [auth]
      summary: Send a magic link
      parameters:
        - $ref: "#/components/parameters/ApiKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Email"
      responses:
        "200":
          description: Always ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ok"
        "429":
          $ref: "#/components/responses/Limited"
  /auth/v1/verify:
    post:
      tags: [auth]
      summary: Consume a magic-link token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Token"
      responses:
        "200":
          description: Session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Error"
  /auth/v1/recover:
    post:
      tags: [auth]
      summary: Send a recovery message
      parameters:
        - $ref: "#/components/parameters/ApiKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Email"
      responses:
        "200":
          description: Always ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ok"
        "429":
          $ref: "#/components/responses/Limited"
  /auth/v1/recover/complete:
    post:
      tags: [auth]
      summary: Set a password from a recovery token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenPassword"
      responses:
        "200":
          description: Session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Error"
  /auth/v1/invite:
    post:
      tags: [auth]
      summary: Invite an email
      security: [{service: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Email"
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ok"
        "409":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Limited"
  /auth/v1/invite/accept:
    post:
      tags: [auth]
      summary: Accept an invite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenPassword"
      responses:
        "200":
          description: Session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Error"
...
components:
...
  parameters:
...
    ApiKey:
      name: Authorization
      in: header
      required: false
      description: Bearer anon key, or the project host
      schema: {type: string}
...
  schemas:
    Email:
      type: object
      required: [email]
      properties:
        email: {type: string}
...
    Token:
      type: object
      required: [token]
      properties:
        token: {type: string}
    TokenPassword:
      type: object
      required: [token, password]
      properties:
        token: {type: string}
        password: {type: string, minLength: 8}
    User:
      type: object
      properties:
        id: {type: string}
        email: {type: string}
    Session:
      type: object
      properties:
        access_token: {type: string}
        refresh_token: {type: string}
        user:
          $ref: "#/components/schemas/User"
    Ok:
      type: object
      properties:
        ok: {type: boolean}
    ErrorBody:
      type: object
      properties:
        error: {type: string}
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Limited:
      description: Too many requests
      headers:
        Retry-After:
          schema: {type: string, example: "60"}
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  securitySchemes:
    operator:
      type: http
      scheme: bearer
      description: Operator token
    service:
      type: http
      scheme: bearer
      description: Project service key
    user:
      type: http
      scheme: bearer
      description: User access token
    project:
      type: http
      scheme: bearer
      description: Anon key, user access token, or service key